+++ src/data(negative-archive)/thumbsupLilyNegAugmented.json
