+++ src/data/closedFistNegative.json
