+++ src/model_archive/heart_two_hands_v0.0.1/data/thumbsup_2024-10-04_21-21-52.json
