+++ src/model_archive/heart_two_hands_v0.0.1/heart_two_hands_v0.0.1.onnx
